Fix unicode in Selendroid
