+++ lib/devices/android/android-controller.js
@@ -1006,77 +1006,4 @@ androidController.openNotifications = function (cb) {
   this.proxy(["openNotification"], cb);
 };
 
-androidController.isIMEActivated = function (cb) {
-  // IME is always activated on Android devices
-  cb(null, {
-    status: status.codes.Success.code,
-    value: true
-  });
-};
-
-androidController.availableIMEEngines = function (cb) {
-  logger.debug('Retrieving available IMEs');
-  this.adb.availableIMEs(function (err, engines) {
-    if (err) return cb(err);
-    logger.debug('Engines: ' + JSON.stringify(engines));
-
-    cb(null, {
-      status: status.codes.Success.code,
-      value: engines
-    });
-  });
-};
-
-androidController.getActiveIMEEngine = function (cb) {
-  logger.debug('Retrieving current default IME');
-  this.adb.defaultIME(function (err, engine) {
-    if (err) return cb(err);
-    cb(null, {
-      status: status.codes.Success.code,
-      value: engine
-    });
-  });
-};
-
-androidController.activateIMEEngine = function (imeId, cb) {
-  logger.debug('Attempting to activate IME \'' + imeId + '\'');
-  this.adb.availableIMEs(function (err, engines) {
-    if (err) return cb(err);
-    if (engines.indexOf(imeId) !== -1) {
-      logger.debug('Found installed IME, attempting to activate.');
-      this.adb.enableIME(imeId, function (err) {
-        if (err) return cb(err);
-        this.adb.setIME(imeId, function (err) {
-          if (err) return cb(err);
-          return cb(null, {
-            status: status.codes.Success.code,
-            value: null
-          });
-        });
-      }.bind(this));
-    } else {
-      logger.debug('IME not found, failing.');
-      return cb(null, {
-        status: status.codes.IMENotAvailable.code,
-        message: 'Unable to find requested IME \'' + imeId + '\'.'
-      });
-    }
-  }.bind(this));
-};
-
-androidController.deactivateIMEEngine = function (cb) {
-  logger.debug('Retrieving current default IME');
-  this.adb.defaultIME(function (err, engine) {
-    if (err) return cb(err);
-    logger.debug('Attempting to deactivate \'' + engine + '\'');
-    this.adb.disableIME(engine, function (err) {
-      if (err) return cb(err);
-      cb(null, {
-        status: status.codes.Success.code,
-        value: null
-      });
-    });
-  }.bind(this));
-};
-
 module.exports = androidController;
