+++ lib/devices/android/selendroid.js
@@ -68,6 +68,8 @@ Selendroid.prototype.init = function () {
     , ['POST', new RegExp('^/wd/hub/session/[^/]+/element/[^/]+/value')]
     , ['GET', new RegExp('^/wd/hub/session/[^/]+/network_connection')]
     , ['POST', new RegExp('^/wd/hub/session/[^/]+/network_connection')]
+    , ['POST', new RegExp('^/wd/hub/session/[^/]+/ime')]
+    , ['GET', new RegExp('^/wd/hub/session/[^/]+/ime')]
   ];
   this.curContext = this.defaultContext();
 };
@@ -238,7 +240,7 @@ Selendroid.prototype.checkSelendroidCerts = function (cb) {
 
 Selendroid.prototype.stop = function (ocb) {
   var completeShutdown = function (cb) {
-    if (this.args.unicodeKeyboard && this.resetKeyboard && this.defaultIME) {
+    if (this.args.unicodeKeyboard && this.args.resetKeyboard && this.defaultIME) {
       logger.debug('Resetting IME to \'' + this.defaultIME + '\'');
       this.adb.setIME(this.defaultIME, function (err) {
         if (err) {
@@ -517,7 +519,8 @@ Selendroid.prototype.setValue = function (elementId, value, cb) {
   logger.debug('Setting text on element \'' + elementId + '\': \'' + value + '\'');
   for (var i = 0; i < value.length; i++) {
     var c = value.charCodeAt(i);
-    if (c > 127 || c === 38) {
+    // if we're using the unicode keyboard, and this is unicode, maybe encode
+    if (this.args.unicodeKeyboard && (c > 127 || c === 38)) {
       // this is not simple ascii, or it is an ampersand (`&`)
       if (c >= parseInt("E000", 16) && c <= parseInt("E040", 16)) {
         // Selenium uses a Unicode PUA to cover certain special characters
