+++ test/functional/selendroid/keyboard-specs.js
@@ -1,0 +1,5 @@
+"use strict";
+
+var selendroidKeyboardTests = require('../common/android-keyboard-base');
+
+describe('selendroid - keyboard', selendroidKeyboardTests);
